# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cluvi API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get current user profile and stats
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /study-packs:
    get:
      operationId: listStudyPacks
      tags: [study-packs]
      summary: List all study packs for current user
      responses:
        "200":
          description: List of study packs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyPack"
    post:
      operationId: createStudyPack
      tags: [study-packs]
      summary: Create a study pack from text content (triggers AI generation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyPackInput"
      responses:
        "201":
          description: Created study pack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPack"
...
  /study-packs/{id}:
    get:
      operationId: getStudyPack
      tags: [study-packs]
      summary: Get a study pack with all generated content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Study pack with content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPackFull"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteStudyPack
      tags: [study-packs]
      summary: Delete a study pack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /study-packs/{id}/generate:
    post:
      operationId: generateStudyPackContent
      tags: [study-packs]
      summary: Re-generate or expand AI content for a study pack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated study pack with regenerated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPackFull"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        displayName:
          type: ["string", "null"]
        isPro:
          type: boolean
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        lastActiveDate:
          type: ["string", "null"]
          format: date
        totalPacks:
          type: integer
        totalFlashcardsStudied:
          type: integer
        totalQuizzesTaken:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - isPro
        - xp
        - level
        - streak
        - totalPacks
        - totalFlashcardsStudied
        - totalQuizzesTaken
        - createdAt
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
...
    StudyPack:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        sourceType:
          type: string
          enum: [text, pdf, image]
        status:
          type: string
          enum: [processing, ready, error]
        summary:
          type: ["string", "null"]
        flashcardCount:
          type: integer
        quizCount:
          type: integer
        examPredictionCount:
          type: integer
        studyStrength:
          type: integer
          description: 0-100 mastery score for this pack
        isPro:
          type: boolean
          description: Whether this pack has pro features unlocked
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - sourceType
        - status
        - flashcardCount
        - quizCount
        - examPredictionCount
        - studyStrength
        - isPro
        - createdAt
        - updatedAt
    StudyPackInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        sourceType:
          type: string
          enum: [text, pdf, image]
        content:
          type: string
          description: Raw text content or base64-encoded file
      required:
        - title
        - sourceType
        - content
    StudyPackFull:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        sourceType:
          type: string
          enum: [text, pdf, image]
        status:
          type: string
          enum: [processing, ready, error]
        summary:
          type: ["string", "null"]
        keyConcepts:
          type: array
          items:
            type: string
        examPredictions:
          type: array
          items:
            type: string
        flashcardCount:
          type: integer
        quizCount:
          type: integer
        examPredictionCount:
          type: integer
        studyStrength:
          type: integer
        isPro:
          type: boolean
        flashcards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
        quizQuestions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - sourceType
        - status
        - flashcardCount
        - quizCount
        - examPredictionCount
        - studyStrength
        - isPro
        - flashcards
        - quizQuestions
        - createdAt
        - updatedAt
...
    Flashcard:
      type: object
      properties:
        id:
          type: integer
        studyPackId:
          type: integer
        front:
          type: string
        back:
          type: string
        confidenceLevel:
          type: integer
          description: 1-5 confidence rating from user
        nextReviewAt:
          type: ["string", "null"]
          format: date-time
        timesStudied:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - studyPackId
        - front
        - back
        - confidenceLevel
        - timesStudied
        - createdAt
...
    QuizQuestion:
      type: object
      properties:
        id:
          type: integer
        studyPackId:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: integer
          description: Index of correct option (0-based)
        explanation:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - studyPackId
        - question
        - options
        - correctAnswer
        - createdAt
...