# lib/api-zod/src/generated/types/studyPack.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cluvi API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudyPackSourceType } from "./studyPackSourceType";
import type { StudyPackStatus } from "./studyPackStatus";

export interface StudyPack {
  id: number;
  userId: number;
  title: string;
  sourceType: StudyPackSourceType;
  status: StudyPackStatus;
  /** @nullable */
  summary?: string | null;
  flashcardCount: number;
  quizCount: number;
  examPredictionCount: number;
  /** 0-100 mastery score for this pack */
  studyStrength: number;
  /** Whether this pack has pro features unlocked */
  isPro: boolean;
  createdAt: Date;
  updatedAt: Date;
}