# lib/api-zod/src/generated/types/studyPackFull.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cluvi API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Flashcard } from "./flashcard";
import type { QuizQuestion } from "./quizQuestion";
import type { StudyPackFullSourceType } from "./studyPackFullSourceType";
import type { StudyPackFullStatus } from "./studyPackFullStatus";

export interface StudyPackFull {
  id: number;
  userId: number;
  title: string;
  sourceType: StudyPackFullSourceType;
  status: StudyPackFullStatus;
  /** @nullable */
  summary?: string | null;
  keyConcepts?: string[];
  examPredictions?: string[];
  flashcardCount: number;
  quizCount: number;
  examPredictionCount: number;
  studyStrength: number;
  isPro: boolean;
  flashcards: Flashcard[];
  quizQuestions: QuizQuestion[];
  createdAt: Date;
  updatedAt: Date;
}