# lib/api-zod/src/generated/types/studyPackInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cluvi API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudyPackInputSourceType } from "./studyPackInputSourceType";

export interface StudyPackInput {
  /** @minLength 1 */
  title: string;
  sourceType: StudyPackInputSourceType;
  /** Raw text content or base64-encoded file */
  content: string;
}