# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cluvi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  clerkId: string;
  /** @nullable */
  displayName?: string | null;
  isPro: boolean;
  xp: number;
  level: number;
  streak: number;
  /** @nullable */
  lastActiveDate?: Date | null;
  totalPacks: number;
  totalFlashcardsStudied: number;
  totalQuizzesTaken: number;
  createdAt: Date;
}